fix: added a history hook
