+++ src/hooks/useFrecency.js
@@ -1,8 +1,5 @@
 import { useEffect, useState } from 'react';
-import { useAuth } from '../contexts/AuthContextProvider';
-import useLocalStorageState from './useLocalStorageState';
-import { db } from '../utils/FirebaseConfig';
-import { query, collection, onSnapshot, orderBy } from 'firebase/firestore';
+import useHistory from './useHistory';
 
 const getPartOfDay = (now) => {
 	const hours = now.getHours();
@@ -25,31 +22,9 @@ const getPartOfWeek = (now) => {
 };
 
 export default function useFrecency() {
-	const { user } = useAuth();
-	const [history, setHistory] = useLocalStorageState('history', []);
+	const history = useHistory();
 	const [suggestions, setSuggestions] = useState([]);
 
-	useEffect(() => {
-		if (!user) return;
-
-		const queryRef = query(collection(db, 'users', `${user.uid}/history`), orderBy('visitTime', 'desc'));
-
-		const unsubscribe = onSnapshot(queryRef, (snapshot) => {
-			let historyArray = [];
-			snapshot.forEach((doc) => {
-				const data = doc.data();
-				const historyData = {
-					visitTime: data.visitTime,
-					title: data.title,
-					url: data.url,
-					favicon: data.favicon,
-				};
-				historyArray.push(historyData);
-			});
-			setHistory(historyArray);
-		});
-		return () => unsubscribe();
-	}, [user, setHistory]);
 
 	useEffect(() => {
 		let now = new Date();
